+++ src/MeasurementFormatting.hpp
@@ -19,7 +19,29 @@ namespace sensirion::upt::mqtt
         }
     };
 
-    struct DefaultMeasurmentToTopicSuffix{
+    struct FullMeasurementFormatter
+    {
+        std::string operator () (const sensirion::upt::core::Measurement& m){
+            std::stringstream stream{};
+            stream.precision(4);
+            stream << "{\"time_offset_ms\":";
+            stream << m.dataPoint.t_offset;
+            stream << ", \"value\":";
+            stream << m.dataPoint.value;
+            stream << ", \"device_id\":";
+            stream << m.metaData.deviceID;
+            stream << ", \"device_type\":";
+            stream << m.metaData.deviceType;
+            stream << ", \"signal\":";
+            stream << quantityOf(m.signalType);
+            stream << ", \"signal_unit\":";
+            stream << unitOf(m.signalType);
+            stream << "}";
+            return stream.str();
+        }
+    };
+
+    struct DefaultMeasurementToTopicSuffix{
         std::string operator() (const sensirion::upt::core::Measurement& m){
             
             std::stringstream stream{};
@@ -27,6 +49,23 @@ namespace sensirion::upt::mqtt
             stream << " " << m.metaData.deviceID 
                    << " " << sensirion::upt::core::quantityOf(m.signalType);
             return stream.str();
+        }
+    };
+    
+    struct MeasurementToTopicSuffixTree{
+        std::string operator() (const sensirion::upt::core::Measurement& m){
+            
+            std::stringstream stream{};
+            stream << core::deviceLabel(m.metaData.deviceType); 
+            stream << "/" << m.metaData.deviceID 
+                   << "/" << sensirion::upt::core::quantityOf(m.signalType);
+            return stream.str();
+        }
+    };
+
+    struct MeasurementToTopicSuffixEmpty{
+        std::string operator() (const sensirion::upt::core::Measurement& m){
+            return "";
         }
     };