+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 0.4.0
+### Added
+- Added MeasurementToTopicSuffixEmpty, to allow more simplistic message handling.
+- Added MeasurementToTopicSuffixTree, to allow hierachical topic handling.
+- Added FullMeasurementFormatter, to go with MeasurementToTopicSuffixEmpty
+
+### Fixed 
+- [BREAKING] Fixed typo in DefaultMeasurementToTopicSuffix.
+
 ## 0.3.0
 ### Changed
 - Moved to stable UPT Core 1.0.0